modif Makevars + gwas in float

+++ src/gwas_approx_pql_bed.cpp
@@ -4,8 +4,8 @@
 
 //[[Rcpp::export]]
 List GWAS_approx_pql_bed(XPtr<matrix4> pA, NumericVector PY, NumericMatrix P, NumericVector p, int beg, int end) {
-  snp_filler_additive_bed<double> S(pA, p, beg, end);
-  gwas_approx_pql<double> x(PY, P, S);
+  snp_filler_additive_bed<float> S(pA, p, beg, end);
+  gwas_approx_pql<float> x(PY, P, S);
   x.run_tests();
   return S.L;
 }
